% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{diff_penalty}
\alias{diff_penalty}
\title{Difference penalty matrix}
\usage{
diff_penalty(K, r = 3L, return_D = FALSE)
}
\arguments{
\item{K}{Length of the coefficient vector (number of B-splines).}

\item{r}{Penalty order (default 3, as recommended by Lambert and
Eilers 2009 because it tends to a normal density at the limit of
strong smoothing).}

\item{return_D}{Logical. If \code{TRUE}, the function returns a list
with components \code{D} (the \eqn{(K-r) \times K} difference
matrix) and \code{P} (the \eqn{K \times K} penalty matrix). If
\code{FALSE} (the default), only \code{P} is returned as a matrix.}
}
\value{
Either a matrix \code{P} or a list (see \code{return_D}).
}
\description{
Returns the matrix \eqn{P = D'D} that gives the sum of squared
\eqn{r}th-order differences of a vector of length \eqn{K}, i.e.
\eqn{\sum_k (\Delta^r \phi_k)^2 = \phi' P \phi}. The matrix \eqn{D}
itself can be returned via \code{return_D = TRUE}.
}
\details{
The rank of \eqn{P} is \eqn{K - r}; its null space is spanned by
polynomials of degree up to \eqn{r-1} (in the sequence index). This
rank-deficiency is the reason for the identifiability constraint
\eqn{\sum_k \phi_k = 0} in the model.
}
\examples{
diff_penalty(K = 6, r = 2)

}
