% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{latent_density}
\alias{latent_density}
\title{Evaluate the latent density on the fine grid}
\usage{
latent_density(phi, basis)
}
\arguments{
\item{phi}{Numeric vector of B-spline coefficients of length \eqn{K}.}

\item{basis}{A \code{"pclm_basis"} object, as returned by
\code{\link{bspline_basis}}.}
}
\value{
Numeric vector of length \eqn{I} (number of fine grid
  intervals), summing to 1.
}
\description{
Returns the discrete latent density values
\eqn{\pi_i = f_Y(u_i) \Delta} at the fine-grid midpoints
\eqn{u_i}, given a vector of B-spline coefficients.
}
\examples{
grid <- seq(0, 10, length.out = 51)
mids <- (head(grid, -1) + tail(grid, -1)) / 2
basis <- bspline_basis(mids, a = 0, b = 10, ndx = 17)
phi <- rep(0, basis$K)
pi_grid <- latent_density(phi, basis)
sum(pi_grid)  # 1

}
